@@ -1,81 +1,71 @@
 {
     "type": "project",
     "license": "proprietary",
+    "minimum-stability": "stable",
+    "prefer-stable": true,
     "require": {
-        "php": "^8.0",
-        "ext-ctype": "*",
-        "ext-fileinfo": "*",
-        "ext-gd": "*",
-        "ext-iconv": "*",
-        "ext-intl": "*",
-        "ext-json": "*",
-        "ext-sodium": "*",
-        "ext-zip": "*",
-        "alterphp/easyadmin-extension-bundle": "^3.1",
-        "api-platform/core": "^2.6",
-        "composer/package-versions-deprecated": "1.11.99.4",
-        "doctrine/annotations": "^1.0",
-        "doctrine/dbal": "^2.0",
-        "doctrine/doctrine-bundle": "^2.4",
-        "doctrine/doctrine-fixtures-bundle": "^3.4",
-        "doctrine/doctrine-migrations-bundle": "^3.2",
-        "doctrine/orm": "^2.10",
-        "easycorp/easyadmin-bundle": "^2.3",
-        "gumlet/php-image-resize": "^1.9",
-        "guzzlehttp/guzzle": "^6.5",
-        "karser/karser-recaptcha3-bundle": "^0.1.17",
-        "konekt/pdf-invoice": "^1.6",
-        "laminas/laminas-code": "^4",
-        "lcobucci/jwt": "~4",
-        "lexik/jwt-authentication-bundle": "^2.11",
-        "nelmio/cors-bundle": "^2.0",
-        "phpdocumentor/reflection-docblock": "^5.3",
-        "sensio/framework-extra-bundle": "^5.5",
-        "smalot/pdfparser": "^0.18.2",
-        "stof/doctrine-extensions-bundle": "^1.5",
-        "symfony/asset": "5.3.*",
-        "symfony/cache": "5.3.*",
-        "symfony/console": "5.3.*",
-        "symfony/dotenv": "5.3.*",
-        "symfony/expression-language": "5.3.*",
-        "symfony/flex": "^1.3.1",
-        "symfony/framework-bundle": "5.3.*",
-        "symfony/google-mailer": "5.3.*",
-        "symfony/mailer": "5.3.*",
-        "symfony/property-access": "5.3.*",
-        "symfony/property-info": "5.3.*",
-        "symfony/proxy-manager-bridge": "5.3.*",
-        "symfony/security-bundle": "5.3.*",
-        "symfony/serializer": "5.3.*",
-        "symfony/templating": "5.3.*",
-        "symfony/twig-bundle": "5.3.*",
-        "symfony/ux-chartjs": "^1.3",
-        "symfony/validator": "5.3.*",
-        "symfony/webpack-encore-bundle": "^1.7",
-        "symfony/yaml": "5.3.*",
+      "php": ">=8.1",
+      "ext-ctype": "*",
+      "ext-iconv": "*",
+      "ext-intl": "*",
+      "doctrine/annotations": "^1.13",
+      "doctrine/doctrine-bundle": "^2.7",
+      "doctrine/orm": "^2.12",
+      "easycorp/easyadmin-bundle": "^4.3",
+      "gumlet/php-image-resize": "^2.0",
+      "karser/karser-recaptcha3-bundle": "^0.1.22",
+      "konekt/pdf-invoice": "^1.12",
+      "php-ds/php-ds": "^1.4",
+      "phpdocumentor/reflection-docblock": "^5.3",
+      "phpstan/phpdoc-parser": "^1.6",
+      "sensio/framework-extra-bundle": "^6.1",
+      "smalot/pdfparser": "^2.2",
+      "symfony/asset": "6.0.*",
+      "symfony/console": "*",
+      "symfony/doctrine-messenger": "6.0.*",
+      "symfony/dotenv": "*",
+      "symfony/expression-language": "6.0.*",
+      "symfony/flex": "^2",
+      "symfony/form": "6.0.*",
+      "symfony/framework-bundle": "*",
+      "symfony/google-mailer": "6.0.*",
+      "symfony/http-client": "6.0.*",
+      "symfony/intl": "6.0.*",
+      "symfony/mailer": "6.0.*",
+      "symfony/mime": "6.0.*",
+      "symfony/monolog-bundle": "^3.0",
+      "symfony/notifier": "6.0.*",
+      "symfony/process": "6.0.*",
+        "symfony/property-access": "6.0.*",
+        "symfony/property-info": "6.0.*",
+        "symfony/proxy-manager-bridge": "6.0.*",
+        "symfony/runtime": "*",
+        "symfony/security-bundle": "6.0.*",
+        "symfony/serializer": "6.0.*",
+        "symfony/string": "6.0.*",
+        "symfony/translation": "6.0.*",
+        "symfony/twig-bundle": "6.0.*",
+        "symfony/ux-chartjs": "^2.2",
+        "symfony/validator": "6.0.*",
+        "symfony/web-link": "6.0.*",
+        "symfony/webpack-encore-bundle": "^1.14",
+        "symfony/yaml": "*",
         "twig/extra-bundle": "^2.12|^3.0",
+        "twig/intl-extra": "^3.4",
         "twig/twig": "^2.12|^3.0",
-        "vich/uploader-bundle": "^1.13"
-    },
-    "require-dev": {
-        "dbrekelmans/bdi": "^0.3.0",
-        "symfony/browser-kit": "^5.2",
-        "symfony/css-selector": "^5.2",
-        "symfony/maker-bundle": "^1.14",
-        "symfony/panther": "^1.0",
-        "symfony/phpunit-bridge": "^5.2",
-        "symfony/stopwatch": "^5.3",
-        "symfony/web-profiler-bundle": "^5.3"
+        "vich/uploader-bundle": "^1.19"
     },
     "config": {
+        "allow-plugins": {
+            "composer/package-versions-deprecated": true,
+            "symfony/flex": true,
+            "symfony/runtime": true
+        },
+        "optimize-autoloader": true,
         "preferred-install": {
             "*": "dist"
         },
-        "sort-packages": true,
-        "allow-plugins": {
-            "composer/package-versions-deprecated": true,
-            "symfony/flex": true
-        }
+        "sort-packages": true
     },
     "autoload": {
         "psr-4": {
@@ -114,8 +107,18 @@
     "extra": {
         "symfony": {
             "allow-contrib": false,
-            "require": "5.3.*",
-            "docker": true
+            "require": "6.0.*"
         }
+    },
+    "require-dev": {
+        "phpunit/phpunit": "^9.5",
+        "rector/rector": "^0.13.7",
+        "symfony/browser-kit": "6.0.*",
+        "symfony/css-selector": "6.0.*",
+        "symfony/debug-bundle": "6.0.*",
+        "symfony/maker-bundle": "^1.0",
+        "symfony/phpunit-bridge": "^6.1",
+        "symfony/stopwatch": "6.0.*",
+        "symfony/web-profiler-bundle": "6.0.*"
     }
 }
