@@ -2,58 +2,45 @@
 
 namespace App\Entity;
 
+use App\Repository\ActivityRepository;
+use DateTimeInterface;
+use Doctrine\DBAL\Types\Types;
 use Doctrine\ORM\Mapping as ORM;
+use Stringable;
 
-/**
- * @ORM\Entity(repositoryClass="App\Repository\ActivityRepository")
- */
-class Activity
+#[ORM\Entity(repositoryClass: ActivityRepository::class)]
+class Activity implements Stringable
 {
-    /**
-     * @ORM\Id()
-     * @ORM\GeneratedValue()
-     * @ORM\Column(type="integer")
-     */
-    private $id;
+    #[ORM\Id]
+    #[ORM\GeneratedValue]
+    #[ORM\Column(type: Types::INTEGER)]
+    private int $id;
 
-    /**
-     * @ORM\Column(type="date")
-     */
-    private $date;
+    #[ORM\Column(type: Types::DATE_MUTABLE)]
+    private ?DateTimeInterface $date = null;
 
-    /**
-     * @ORM\Column(type="decimal", precision=3, scale=1, nullable=true)
-     */
-    private $value;
+    #[ORM\Column(type: Types::DECIMAL, precision: 3, scale: 1, nullable: true)]
+    private ?string $value = null;
 
-    /**
-     * @ORM\ManyToOne(targetEntity="App\Entity\Company", inversedBy="activities", cascade={"persist"})
-     */
-    private $company;
+    #[ORM\ManyToOne(targetEntity: Company::class, cascade: ['persist'], inversedBy: 'activities')]
+    private ?Company $company = null;
 
-    /**
-     * @ORM\ManyToOne(targetEntity="App\Entity\Invoice", inversedBy="activities")
-     */
-    private $invoice;
+    #[ORM\ManyToOne(targetEntity: Invoice::class, inversedBy: 'activities')]
+    private ?Invoice $invoice = null;
 
-    public function __toString()
+    public function __toString(): string
     {
-        return $this->getCompany() . ' - ' . $this->getValue() . ' - '. $this->getDate()->format('d/m/Y');
+        return $this->getCompany() . ' - ' . $this->getValue() . ' - ' . $this->getDate()->format('d/m/Y');
     }
 
-    public function getId(): ?int
+    public function getCompany(): ?Company
     {
-        return $this->id;
+        return $this->company;
     }
 
-    public function getDate(): ?\DateTimeInterface
+    public function setCompany(?Company $Company): self
     {
-        return $this->date;
-    }
-
-    public function setDate(\DateTimeInterface $date): self
-    {
-        $this->date = $date;
+        $this->company = $Company;
 
         return $this;
     }
@@ -70,16 +57,21 @@ class Activity
         return $this;
     }
 
-    public function getCompany(): ?Company
+    public function getDate(): ?DateTimeInterface
     {
-        return $this->company;
+        return $this->date;
     }
 
-    public function setCompany(?Company $Company): self
+    public function setDate(DateTimeInterface $date): self
     {
-        $this->company = $Company;
+        $this->date = $date;
 
         return $this;
+    }
+
+    public function getId(): ?int
+    {
+        return $this->id;
     }
 
     public function getInvoice(): ?Invoice
