@@ -2,86 +2,77 @@
 
 namespace App\Entity;
 
-use App\Helper\StringHelper;
+use App\Repository\ExperienceRepository;
+use DateTime;
+use DateTimeInterface;
 use Doctrine\Common\Collections\ArrayCollection;
 use Doctrine\Common\Collections\Collection;
+use Doctrine\DBAL\Types\Types;
 use Doctrine\ORM\Mapping as ORM;
-use Symfony\Bridge\Doctrine\Validator\Constraints\UniqueEntity;
+use Stringable;
 
-/**
- * @ORM\Entity(repositoryClass="App\Repository\ExperienceRepository")
- */
-class Experience
+#[ORM\Entity(repositoryClass: ExperienceRepository::class)]
+class Experience implements Stringable
 {
-    /**
-     * @ORM\Id()
-     * @ORM\GeneratedValue()
-     * @ORM\Column(type="integer")
-     */
-    private $id;
+    #[ORM\Id]
+    #[ORM\GeneratedValue]
+    #[ORM\Column(type: Types::INTEGER)]
+    private int $id;
+
+    #[ORM\Column(type: Types::STRING, length: 255)]
+    private string $title;
+
+    #[ORM\Column(type: Types::STRING, length: 255, nullable: true)]
+    private ?string $search = null;
+
+    #[ORM\Column(type: Types::TEXT, nullable: true)]
+    private ?string $description = null;
+
+    #[ORM\Column(type: Types::BOOLEAN)]
+    private bool $isFreelance;
+
+    #[ORM\Column(type: Types::BOOLEAN)]
+    private bool $onSite;
+
+    #[ORM\Column(type: Types::BOOLEAN)]
+    private bool $onHomepage;
+
+    #[ORM\Column(type: Types::DATE_MUTABLE)]
+    private DateTimeInterface $dateBegin;
+
+    #[ORM\Column(type: Types::DATE_MUTABLE, nullable: true)]
+    private ?DateTimeInterface $dateEnd = null;
 
     /**
-     * @ORM\Column(type="string", length=255)
+     * @var Collection<Skill>
      */
-    private $title;
+    #[ORM\ManyToMany(targetEntity: Skill::class, inversedBy: 'experiences', cascade: ['persist'])]
+    private Collection $skills;
+
+    #[ORM\ManyToOne(targetEntity: Company::class, cascade: ['persist'], inversedBy: 'experiences')]
+    #[ORM\JoinColumn(nullable: false)]
+    private ?Company $company = null;
 
     /**
-     * @ORM\Column(type="string", length=255, nullable=true)
+     * @var Collection<Invoice>
      */
-    private $search;
+    #[ORM\OneToMany(mappedBy: 'experience', targetEntity: Invoice::class, cascade: ['persist'])]
+    private Collection $invoices;
 
-    /**
-     * @ORM\Column(type="text", nullable=true)
-     */
-    private $description;
+    #[ORM\ManyToOne(targetEntity: Company::class)]
+    #[ORM\JoinColumn(nullable: true)]
+    private ?Company $client = null;
 
-    /**
-     * @ORM\Column(type="boolean")
-     */
-    private $isFreelance;
-
-    /**
-     * @ORM\Column(type="boolean")
-     */
-    private $onSite;
-
-    /**
-     * @ORM\Column(type="boolean")
-     */
-    private $onHomepage;
-
-    /**
-     * @ORM\Column(type="date")
-     */
-    private $dateBegin;
-
-    /**
-     * @ORM\Column(type="date", nullable=true)
-     */
-    private $dateEnd;
-
-    /**
-     * @ORM\ManyToMany(targetEntity="App\Entity\Skill", inversedBy="experiences", cascade={"persist"})
-     */
-    private $skills;
-
-    /**
-     * @ORM\ManyToOne(targetEntity="App\Entity\Company", inversedBy="experiences", cascade={"persist"})
-     * @ORM\JoinColumn(nullable=false)
-     */
-    private $company;
-
-    /**
-     * @ORM\OneToMany(targetEntity="App\Entity\Invoice", mappedBy="experience", cascade={"persist"})
-     */
-    private $invoices;
-
-    /**
-     * @var Company
-     * @ORM\ManyToOne(targetEntity="App\Entity\Company")
-     * @ORM\JoinColumn(nullable=true)
-     */
-    private $client;
+    public function __construct()
+    {
+        $this->isFreelance = true;
+        $this->onSite = true;
+        $this->onHomepage = true;
+        $this->dateBegin = new DateTime();
+        $this->title = "Développeur Web";
+        $this->skills = new ArrayCollection();
+        $this->invoices = new ArrayCollection();
+    }
 
     public function __toString(): string
     {
@@ -95,23 +86,49 @@ class Experience
         return $str;
     }
 
-    public function __construct()
+    public function getCompany(): ?Company
     {
-        $this->skills = new ArrayCollection();
-        $this->isFreelance = true;
-        $this->onSite = true;
-        $this->onHomepage = true;
-        $this->dateBegin = new \DateTime();
-        $this->title = "Développeur Web";
-        $this->invoices = new ArrayCollection();
+        return $this->company;
+    }
+
+    public function setCompany(?Company $company): self
+    {
+        $this->company = $company;
+        $this->client = $company;
+
+        return $this;
+    }
+
+    public function getDateBegin(): ?DateTimeInterface
+    {
+        return $this->dateBegin;
+    }
+
+    public function setDateBegin(DateTimeInterface $dateBegin): self
+    {
+        $this->dateBegin = $dateBegin;
+
+        return $this;
+    }
+
+    public function getDateEnd(): ?DateTimeInterface
+    {
+        return $this->dateEnd;
+    }
+
+    public function setDateEnd(?DateTimeInterface $dateEnd): self
+    {
+        $this->dateEnd = $dateEnd;
+
+        return $this;
     }
 
     public function getMainSkills(): string
     {
         /** @var Skill[] $skillCollection */
         $skillCollection = $this->getSkills()->filter(
-            function(Skill $var) {
-                if (substr($var->getName(), 0, 7) === "Symfony" || substr($var->getName(), 0, 7) === "Angular") {
+            function (Skill $var) {
+                if (str_starts_with($var->getName(), "Symfony") || str_starts_with($var->getName(), "Angular")) {
                     return true;
                 }
                 return false;
@@ -127,7 +144,15 @@ class Experience
         return implode(', ', $skillNames);
     }
 
-    public function getId()
+    /**
+     * @return Collection<Skill>
+     */
+    public function getSkills(): Collection
+    {
+        return $this->skills;
+    }
+
+    public function getId(): int
     {
         return $this->id;
     }
@@ -192,38 +217,6 @@ class Experience
         return $this;
     }
 
-    public function getDateBegin(): ?\DateTimeInterface
-    {
-        return $this->dateBegin;
-    }
-
-    public function setDateBegin(\DateTimeInterface $dateBegin): self
-    {
-        $this->dateBegin = $dateBegin;
-
-        return $this;
-    }
-
-    public function getDateEnd(): ?\DateTimeInterface
-    {
-        return $this->dateEnd;
-    }
-
-    public function setDateEnd(?\DateTimeInterface $dateEnd): self
-    {
-        $this->dateEnd = $dateEnd;
-
-        return $this;
-    }
-
-    /**
-     * @return Collection|Skill[]
-     */
-    public function getSkills(): Collection
-    {
-        return $this->skills;
-    }
-
     public function addSkill(Skill $skill): self
     {
         if (!$this->skills->contains($skill)) {
@@ -254,21 +247,8 @@ class Experience
         return $this;
     }
 
-    public function getCompany(): ?Company
-    {
-        return $this->company;
-    }
-
-    public function setCompany(?Company $company): self
-    {
-        $this->company = $company;
-        $this->client = $company;
-
-        return $this;
-    }
-
     /**
-     * @return Collection|Invoice[]
+     * @return Collection<Invoice>
      */
     public function getInvoices(): Collection
     {
@@ -303,16 +283,15 @@ class Experience
         return $this->client;
     }
 
-
-    public function getClientName(): string
-    {
-        return $this->client ? $this->client->getName() : '';
-    }
-
     public function setClient(?Company $client): self
     {
         $this->client = $client;
 
         return $this;
+    }
+
+    public function getClientName(): string
+    {
+        return $this->client ? $this->client->getName() : '';
     }
 }
