@@ -3,63 +3,72 @@
 namespace App\Entity;
 
 use App\Repository\StatementRepository;
+use DateTime;
+use DateTimeInterface;
+use Doctrine\DBAL\Types\Types;
 use Doctrine\ORM\Mapping as ORM;
-use Symfony\Component\HttpFoundation\File\File;
-use Vich\UploaderBundle\Mapping\Annotation as Vich;
 use Symfony\Bridge\Doctrine\Validator\Constraints\UniqueEntity;
+use Symfony\Component\HttpFoundation\File\File;
+use Symfony\Contracts\Translation\TranslatorInterface;
+use Vich\UploaderBundle\Mapping\Annotation as Vich;
 
 /**
- * @UniqueEntity(fields={"filename"}, message="Operation already exists")
- * @ORM\Entity(repositoryClass=StatementRepository::class)
  * @Vich\Uploadable
  */
+#[UniqueEntity(fields: ['filename'], message: 'Operation already exists')]
+#[ORM\Entity(repositoryClass: StatementRepository::class)]
 class Statement
 {
-    /**
-     * @ORM\Id
-     * @ORM\GeneratedValue
-     * @ORM\Column(type="integer")
-     */
-    private $id;
+    #[ORM\Id]
+    #[ORM\GeneratedValue]
+    #[ORM\Column(type: Types::INTEGER)]
+    private int $id;
 
-    /**
-     * @ORM\Column(type="date", nullable=true)
-     */
-    private $date;
+    #[ORM\Column(type: Types::DATE_MUTABLE, nullable: true)]
+    private ?DateTimeInterface $date = null;
 
-    /**
-     * @ORM\Column(type="string", length=255, unique=true)
-     */
-    private $filename;
+    #[ORM\Column(type: Types::STRING, length: 255, unique: true)]
+    private ?string $filename = null;
 
-    /**
-     * @ORM\Column(type="datetime", nullable=true)
-     * @var \DateTime
-     */
-    private $updatedAt;
+    #[ORM\Column(type: 'datetime', nullable: true)]
+    private ?DateTimeInterface $updatedAt = null;
 
     /**
      * @Vich\UploadableField(mapping="statements", fileNameProperty="filename")
-     * @var File
      */
-    private $file;
+    private ?File $file = null;
 
-    /**
-     * @ORM\Column(type="integer", nullable=true)
-     */
-    private $operationsCount;
+    #[ORM\Column(type: Types::INTEGER, nullable: true)]
+    private ?int $operationsCount = null;
+
+    private ?TranslatorInterface $translator = null;
+
+    public function setTranslator(TranslatorInterface $translator): void
+    {
+        $this->translator = $translator;
+    }
+
+    public function __toString(): string
+    {
+        $date = $this->getDate();
+        if ($this->translator) {
+            return $this->translator->trans($date->format('F')) . " " . $date->format('Y');
+        } else {
+            return $date->format('m') . " " . $date->format('Y');
+        }
+    }
 
     public function getId(): ?int
     {
         return $this->id;
     }
 
-    public function getDate(): ?\DateTimeInterface
+    public function getDate(): ?DateTimeInterface
     {
         return $this->date;
     }
 
-    public function setDate(?\DateTimeInterface $date): self
+    public function setDate(?DateTimeInterface $date): self
     {
         $this->date = $date;
 
@@ -78,7 +87,12 @@ class Statement
         return $this;
     }
 
-    public function setFile(File $file = null)
+    public function getFile(): ?File
+    {
+        return $this->file;
+    }
+
+    public function setFile(File $file = null): void
     {
         $this->file = $file;
 
@@ -87,18 +101,13 @@ class Statement
         // otherwise the event listeners won't be called and the file is lost
         if ($file) {
             // if 'updatedAt' is not defined in your entity, use another property
-            $this->updatedAt = new \DateTime('now');
+            $this->updatedAt = new DateTime('now');
         }
     }
 
-    public function getFile()
+    public function getOperationsCount(): int
     {
-        return $this->file;
-    }
-
-    public function getOperationsCount(): ?int
-    {
-        return $this->operationsCount;
+        return $this->operationsCount ?? 0;
     }
 
     public function setOperationsCount(?int $operationsCount): self
