@@ -4,12 +4,9 @@ namespace App\Form\Type;
 
 use App\Entity\Activity;
 use App\Entity\Company;
-use App\Entity\Invoice;
-use App\Repository\InvoiceRepository;
 use App\Service\ReportService;
-use DateInterval;
-use Doctrine\ORM\EntityRepository;
-use Symfony\Bridge\Doctrine\Form\Type\EntityType;
+use DateTime;
+use Exception;
 use Symfony\Component\Form\AbstractType;
 use Symfony\Component\Form\Extension\Core\Type\CollectionType;
 use Symfony\Component\Form\Extension\Core\Type\SubmitType;
@@ -18,17 +15,18 @@ use Symfony\Component\OptionsResolver\OptionsResolver;
 
 class MonthActivitiesType extends AbstractType
 {
-    /** @var \DateTime $currentDate */
-    public $currentDate;
+    public DateTime $currentDate;
 
-    /** @var ReportService */
-    public $reportService;
+    public ReportService $reportService;
 
     public function __construct(ReportService $reportService)
     {
         $this->reportService = $reportService;
     }
 
+    /**
+     * @throws Exception
+     */
     public function buildForm(FormBuilderInterface $builder, array $options)
     {
         $this->currentDate = $options['currentDate'];
@@ -43,22 +41,21 @@ class MonthActivitiesType extends AbstractType
         $builder
             ->add('activities', CollectionType::class, [
                 'entry_type' => ActivityType::class,
-                'data' => array_values($activitiesData)
+                'data'       => array_values($activitiesData)
             ])
             ->add('send', SubmitType::class, [
                 'attr' => [
                     'class' => 'btn btn-primary action-save'
                 ]
-            ])
-        ;
+            ]);
     }
 
     public function configureOptions(OptionsResolver $resolver)
     {
         $resolver->setDefaults([
-            'activities' => [],
-            'currentDate' => null,
-            'company' => null
-        ]);
+                                   'activities'  => [],
+                                   'currentDate' => null,
+                                   'company'     => null
+                               ]);
     }
 }
