@@ -1,4 +1,4 @@
-import {Controller } from 'stimulus';
+import { Controller } from '@hotwired/stimulus';
 
 /*
  * This is an example Stimulus controller!
@@ -9,8 +9,6 @@ import {Controller } from 'stimulus';
  *
  * Delete this file or adapt it for your use!
  */
-/* webpackMode: 'lazy' */
-/* stimulusFetch: 'lazy' */
 export default class extends Controller {
     connect() {
         this.element.textContent = 'Hello Stimulus! Edit me in assets/controllers/hello_controller.js';
