@@ -2,42 +2,27 @@
 
 namespace App\Command;
 
-use App\Service\CompanyService;
-use App\Service\DeclarationService;
-use App\Service\InvoiceService;
-use App\Service\ReportService;
 use Symfony\Component\Console\Command\Command;
-use Symfony\Component\Console\Input\InputArgument;
 use Symfony\Component\Console\Input\InputInterface;
-use Symfony\Component\Console\Input\InputOption;
 use Symfony\Component\Console\Output\OutputInterface;
 use Symfony\Component\Console\Style\SymfonyStyle;
 use Symfony\Component\DependencyInjection\ParameterBag\ParameterBagInterface;
 use Symfony\Component\Finder\Finder;
 use Symfony\Component\HttpKernel\KernelInterface;
+use Symfony\Component\Mailer\Exception\TransportExceptionInterface;
 use Symfony\Component\Mailer\MailerInterface;
 use Symfony\Component\Mime\Email;
-use Symfony\Component\Templating\EngineInterface;
-use Symfony\Contracts\Translation\TranslatorInterface;
-use Twig\Environment;
 
 class BackupCommand extends Command
 {
     protected static $defaultName = 'app:backup';
 
-    protected $appKernel;
-    protected $params;
-    protected $mailer;
-
     public function __construct(
-        KernelInterface $appKernel,
-        ParameterBagInterface $params,
-        MailerInterface $mailer,
+        protected KernelInterface $appKernel,
+        protected ParameterBagInterface $params,
+        protected MailerInterface $mailer,
     ) {
         parent::__construct();
-        $this->mailer = $mailer;
-        $this->params = $params;
-        $this->appKernel = $appKernel;
     }
 
     protected function configure()
@@ -47,6 +32,9 @@ class BackupCommand extends Command
         ;
     }
 
+    /**
+     * @throws TransportExceptionInterface
+     */
     protected function execute(InputInterface $input, OutputInterface $output): int
     {
         $io = new SymfonyStyle($input, $output);
@@ -74,7 +62,7 @@ class BackupCommand extends Command
                 ->subject($this->params->get('MAILER_SUBJECT') . ' Backup')
                 ->attachFromPath(
                     $backupFile,
-                    $backupName);;
+                    $backupName);
 
             $this->mailer->send($email);
 
