@@ -1,5 +1,5 @@
 <?php
-
+// src/Controller/IndexController.php
 namespace App\Controller;
 
 use App\Form\Type\ContactFormType;
@@ -9,65 +9,45 @@ use App\Repository\ExperienceRepository;
 use App\Repository\HobbyRepository;
 use App\Repository\LinkRepository;
 use App\Repository\SkillRepository;
+use Doctrine\ORM\Query\QueryException;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
-use Symfony\Component\Asset\Packages;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
-use Symfony\Component\HttpKernel\KernelInterface;
+use Symfony\Component\Mailer\Exception\TransportExceptionInterface;
 use Symfony\Component\Mailer\MailerInterface;
 use Symfony\Component\Mime\Email;
 use Symfony\Component\Routing\Annotation\Route;
-use Symfony\Component\Routing\Generator\UrlGeneratorInterface;
 use Symfony\Contracts\Translation\TranslatorInterface;
-use Smalot\PdfParser\Parser;
 
 class IndexController extends AbstractController
 {
     /**
-     * @Route("/", name="index")
-     * @param Request $request
-     * @param AttributeRepository $attributeRepository
-     * @param SkillRepository $skillRepository
-     * @param ExperienceRepository $experienceRepository
-     * @param EducationRepository $educationRepository
-     * @param HobbyRepository $hobbyRepository
-     * @param LinkRepository $linkRepository
-     * @param MailerInterface $mailer
-     * @param TranslatorInterface $translator
-     * @return Response
-     * @throws \Symfony\Component\Mailer\Exception\TransportExceptionInterface
+     * @throws QueryException|TransportExceptionInterface
      */
-    public function index(
-        Request $request,
-        AttributeRepository $attributeRepository,
-        SkillRepository $skillRepository,
-        ExperienceRepository $experienceRepository,
-        EducationRepository $educationRepository,
-        HobbyRepository $hobbyRepository,
-        LinkRepository $linkRepository,
-        MailerInterface $mailer,
-        TranslatorInterface $translator
-    ) {
+    #[Route(path: '/', name: 'app_index')]
+    public function index(Request              $request, AttributeRepository $attributeRepository,
+                          SkillRepository      $skillRepository, ExperienceRepository $experienceRepository,
+                          EducationRepository  $educationRepository, HobbyRepository $hobbyRepository,
+                          LinkRepository       $linkRepository, MailerInterface $mailer, TranslatorInterface $translator
+    ): Response {
         $form = $this->createForm(ContactFormType::class);
         $form->handleRequest($request);
-
         $format = $request->query->get('format');
         $all = $request->query->get('all');
         $experiencesFilter = $all ? [] : ['onHomepage' => true];
-
         $data = [
-            'attributes' => $attributeRepository->findAllIndexedBy('slug', false),
+            'attributes'          => $attributeRepository->findAllIndexedBy('slug', false),
             'attributes_listable' => $attributeRepository->findAllIndexedBy('slug', true),
-            'skills' => $skillRepository->findBy(['onHomepage' => true], ['level' => 'DESC']),
-            'experiences' => $experienceRepository->findBy($experiencesFilter, ['dateBegin' => 'DESC']),
-            'educations' => $educationRepository->findBy([], ['dateBegin' => 'DESC']),
-            'hobbies' => $hobbyRepository->findAll(),
-            'links' => $linkRepository->findAll(),
-            'format' => $format,
-            'all' => $all,
-            'contactForm' => $form->createView(),
-            'filename' => 'jeremy-achain-cv',
-            'messageSended' => $request->get('messageSended')
+            'skills'              => $skillRepository->findBy(['onHomepage' => true], ['level' => 'DESC']),
+            'experiences'         => $experienceRepository->findBy($experiencesFilter, ['dateBegin' => 'DESC']),
+            'educations'          => $educationRepository->findBy([], ['dateBegin' => 'DESC']),
+            'hobbies'             => $hobbyRepository->findAll(),
+            'links'               => $linkRepository->findAll(),
+            'format'              => $format,
+            'all'                 => $all,
+            'contactForm'         => $form->createView(),
+            'filename'            => 'jeremy-achain-cv',
+            'messageSended'       => $request->get('messageSended')
         ];
 
         if ($form->isSubmitted() && $form->isValid()) {
@@ -82,11 +62,9 @@ class IndexController extends AbstractController
 
             $mailer->send($email);
 
-            return $this->redirectToRoute('index', ['messageSended' => true]);
+            return $this->redirectToRoute('app_index', ['messageSended' => true]);
         }
-
         $data['isSubmittedWithErrors'] = $form->isSubmitted() && !$form->isValid();
-
         return $this->render('page/index.html.twig', $data);
     }
 }
