@@ -1,20 +1,24 @@
 <?php
+
 namespace App\DQL;
 
+use Doctrine\ORM\Query\AST\Functions\FunctionNode;
+use Doctrine\ORM\Query\AST\InputParameter;
 use Doctrine\ORM\Query\Lexer;
 use Doctrine\ORM\Query\Parser;
-use Doctrine\ORM\Query\AST\Functions\FunctionNode;
 use Doctrine\ORM\Query\SqlWalker;
-use Doctrine\ORM\Query\AST\InputParameter;
 
 class DistanceFunction extends FunctionNode
 {
     /** @var InputParameter longitudeColumn */
     public $longitudeColumn = null;
+
     /** @var InputParameter latitudeColumn */
     public $latitudeColumn = null;
+
     /** @var InputParameter longitudeValue */
     public $longitudeValue = null;
+
     /** @var InputParameter latitudeValue */
     public $latitudeValue = null;
 
@@ -36,12 +40,12 @@ class DistanceFunction extends FunctionNode
     {
         return 'st_distance(' .
             'point(' .
-                $this->longitudeColumn->dispatch($sqlWalker) . ', ' .
-                $this->latitudeColumn->dispatch($sqlWalker) .
+            $this->longitudeColumn->dispatch($sqlWalker) . ', ' .
+            $this->latitudeColumn->dispatch($sqlWalker) .
             '), point(' .
-                $this->longitudeValue->dispatch($sqlWalker) . ', ' .
-                $this->latitudeValue->dispatch($sqlWalker) .
+            $this->longitudeValue->dispatch($sqlWalker) . ', ' .
+            $this->latitudeValue->dispatch($sqlWalker) .
             ')' .
-        ')';
+            ')';
     }
 }
