@@ -1,18 +1,21 @@
 <?php
+
 namespace App\DQL;
 
+use Doctrine\ORM\Query\AST\Functions\FunctionNode;
+use Doctrine\ORM\Query\AST\InputParameter;
 use Doctrine\ORM\Query\Lexer;
 use Doctrine\ORM\Query\Parser;
-use Doctrine\ORM\Query\AST\Functions\FunctionNode;
 use Doctrine\ORM\Query\SqlWalker;
-use Doctrine\ORM\Query\AST\InputParameter;
 
 class IfFunction extends FunctionNode
 {
     /** @var InputParameter firstValue */
     public $firstValue = null;
+
     /** @var InputParameter secondValue */
     public $secondValue = null;
+
     /** @var InputParameter thirdValue */
     public $thirdValue = null;
 
@@ -34,6 +37,6 @@ class IfFunction extends FunctionNode
             $this->firstValue->dispatch($sqlWalker) . ', ' .
             $this->secondValue->dispatch($sqlWalker) . ', ' .
             $this->thirdValue->dispatch($sqlWalker) .
-        ')';
+            ')';
     }
 }
