@@ -2,14 +2,14 @@
 
 namespace App\Form\Type;
 
+use Karser\Recaptcha3Bundle\Form\Recaptcha3Type;
+use Karser\Recaptcha3Bundle\Validator\Constraints\Recaptcha3;
 use Symfony\Component\Form\AbstractType;
 use Symfony\Component\Form\Extension\Core\Type\EmailType;
 use Symfony\Component\Form\Extension\Core\Type\SubmitType;
 use Symfony\Component\Form\Extension\Core\Type\TextareaType;
 use Symfony\Component\Form\FormBuilderInterface;
 use Symfony\Component\OptionsResolver\OptionsResolver;
-use Karser\Recaptcha3Bundle\Form\Recaptcha3Type;
-use Karser\Recaptcha3Bundle\Validator\Constraints\Recaptcha3;
 
 class ContactFormType extends AbstractType
 {
@@ -23,7 +23,6 @@ class ContactFormType extends AbstractType
                 'constraints' => new Recaptcha3(),
                 'action_name' => 'contact',
             ]);
-        ;
     }
 
     public function configureOptions(OptionsResolver $resolver)
