@@ -3,103 +3,50 @@
 namespace App\Service;
 
 use App\Entity\Declaration;
+use App\Enum\DeclarationTypeEnum;
 use App\Repository\DeclarationRepository;
-use App\Repository\PeriodRepository;
+use DateInterval;
+use DateTime;
 use Doctrine\ORM\EntityManagerInterface;
+use Exception;
 
 class DeclarationTypeService
 {
-    /** @var EntityManagerInterface */
-    private $entityManager;
-
-    /** @var DeclarationRepository */
-    private $declarationRepository;
-
-    /** @var PeriodService */
-    private $periodService;
-
-    /** @var string  */
-    private $type = '';
-
     /**
      * DeclarationTypeService constructor.
      * @param EntityManagerInterface $entityManager
      * @param DeclarationRepository $declarationRepository
      * @param PeriodService $periodService
-     * @param $type
-     * @param $dueSocialMonth
+     * @param DeclarationTypeEnum $type
      */
     public function __construct(
-        EntityManagerInterface $entityManager,
-        DeclarationRepository $declarationRepository,
-        PeriodService $periodService,
-        $type
+        private readonly EntityManagerInterface $entityManager,
+        private readonly DeclarationRepository  $declarationRepository, private readonly PeriodService $periodService,
+        private readonly DeclarationTypeEnum    $type
     ) {
-        $this->entityManager = $entityManager;
-        $this->declarationRepository = $declarationRepository;
-        $this->periodService = $periodService;
-        $this->type = $type;
     }
 
-    public function getDueMonth()
-    {
-        switch ($this->type) {
-            case Declaration::TYPE_SOCIAL:
-                return [
-                    4,
-                    7,
-                    10,
-                    1
-                ];
-                break;
-
-            case Declaration::TYPE_TVA:
-                return [
-                    6,// 7
-                    11,// 12
-                    3 // 4
-                ];
-                break;
-
-            case Declaration::TYPE_IMPOT:
-                return [
-                    5
-                ];
-                break;
-
-
-            case Declaration::TYPE_CFE:
-                return [
-                    11
-                ];
-                break;
-        }
-
-        return [];
-    }
-
-
-    public function getTotalByYear($year)
+    public function getTotalByYear($year): int|string|null
     {
         $declarations = [];
         $amount = 0;
 
         $period = $this->periodService->getAnnualyByYear($year);
         $declaration = $this->declarationRepository->findOneBy([
-            'type' => $this->type,
-            'period' => $period
-        ]);
+                                                                   'type'   => $this->type,
+                                                                   'period' => $period
+                                                               ]);
         if ($declaration) {
             $declarations[] = $declaration;
         }
 
-        if ($this->type === Declaration::TYPE_SOCIAL) {
-            $period = $this->periodService->getQuarterlyByYear($year);
-            foreach ($period as $period) {
+        if ($this->type === DeclarationTypeEnum::Social) {
+            $periods = $this->periodService->getQuarterlyByYear($year);
+            foreach ($periods as $period) {
                 $declaration = $this->declarationRepository->findOneBy([
-                    'type' => $this->type,
-                    'period' => $period
-                ]);
+                                                                           'type'   => $this->type,
+                                                                           'period' => $period
+                                                                       ]);
 
                 if ($declaration) {
                     $declarations[] = $declaration;
@@ -108,36 +55,91 @@ class DeclarationTypeService
         }
 
         foreach ($declarations as $declaration) {
-            $amount+= $declaration->getTax();
+            $amount += $declaration->getTax();
         }
 
         return $amount;
     }
 
     /**
-     * Verifie si on est dans un mois de déclaration
-     * @return bool
+     * Retourne les dates de début et fin du prochain trimestre de cotisation
+     * @param DateTime|null $date
+     * @throws Exception
      */
-    public function isDueMonth()
+    public function getNextDueDate(DateTime $date = null): array
     {
-        $date = new \DateTime();
-        $dueDatesMonth = $this->getDueMonth();
+        $currentDate = $date ?: new DateTime();
+        $dueDates = $this->getDueDatesBy($currentDate);
 
-        foreach ($dueDatesMonth as $index => $dueDateMonth) {
-            if (intval($date->format('m')) ===  $dueDateMonth) {
-                return true;
+        foreach ($dueDates as $index => $dueDate) {
+            if ($currentDate < $dueDate[2]) {
+                $dueDate[] = $currentDate >= $dueDate[1] && $currentDate <= $dueDate[2];
+                $dueDate[] = $this->type;
+                return $dueDate;
             }
         }
 
-        return false;
+        return [];
+    }
+
+    /**
+     * Récupère les dates courantes de déclarations
+     * @throws Exception
+     */
+    public function getDueDatesBy(DateTime $date): array
+    {
+        $dueDatesMonth = $this->getDueMonth();
+        $dueDates = [];
+
+        foreach ($dueDatesMonth as $index => $dueDateMonth) {
+            $dueDateBegin = new DateTime(
+                $this->getAccountingYear($date) . '-' . ($dueDateMonth < 10 ? '0' : '') . $dueDateMonth . '-01'
+            );
+            if ($index == count($dueDatesMonth) - 1) {
+                $dueDateBegin->add(new DateInterval('P1Y'));
+            }
+
+            $dueDates[] = [
+                $index + 1,
+                clone $dueDateBegin,
+                $dueDateBegin
+                    ->add(new DateInterval('P' . (intval($dueDateBegin->format('t')) - 1) . 'D'))
+                    ->add(new DateInterval('PT23H59M59S'))
+            ];
+        }
+
+        return $dueDates;
+    }
+
+    public function getDueMonth(): array
+    {
+        return match ($this->type) {
+            DeclarationTypeEnum::Social => [
+                4,
+                7,
+                10,
+                1
+            ],
+            DeclarationTypeEnum::TVA => [
+                6,
+                11,
+                3
+            ],
+            DeclarationTypeEnum::Impot => [
+                5
+            ],
+            DeclarationTypeEnum::CFE => [
+                11
+            ],
+            default => [],
+        };
+
     }
 
     /**
      * Retourne l'année comptable
-     * @param \DateTime $date
-     * @return int
      */
-    public function getAccountingYear(\DateTime $date): int
+    public function getAccountingYear(DateTime $date): int
     {
         $dueMonth = $this->getDueMonth();
         $lastMonth = array_pop($dueMonth);
@@ -152,73 +154,21 @@ class DeclarationTypeService
     }
 
     /**
-     * Récupère les dates courantes de déclarations
-     * @param \DateTime $date
-     * @return array
-     * @throws \Exception
-     */
-    public function getDueDatesBy(\DateTime $date): array
-    {
-        $dueDatesMonth = $this->getDueMonth();
-        $dueDates = [];
-
-        foreach ($dueDatesMonth as $index => $dueDateMonth) {
-            $dueDateBegin = new \DateTime($this->getAccountingYear($date).'-'.($dueDateMonth < 10 ? '0' : '').$dueDateMonth.'-01');
-            if ($index == count($dueDatesMonth) - 1) {
-                $dueDateBegin->add(new \DateInterval('P1Y'));
-            }
-
-            $dueDates[] = [
-                $index + 1,
-                clone $dueDateBegin,
-                $dueDateBegin
-                    ->add(new \DateInterval('P'.(intval($dueDateBegin->format('t')) - 1).'D'))
-                    ->add(new \DateInterval('PT23H59M59S'))
-            ];
-        }
-
-        return $dueDates;
-    }
-
-    /**
-     * Retourne les dates de début et fin du prochain trimestre de cotisation
-     * @param \DateTime|null $date
-     * @return array
-     * @throws \Exception
-     */
-    public function getNextDueDate(\DateTime $date = null): array
-    {
-        $currentDate = $date ? $date : new \DateTime();
-        $dueDates = $this->getDueDatesBy($currentDate);
-
-        foreach ($dueDates as $index => $dueDate)
-        {
-            if ($currentDate < $dueDate[2]) {
-                $dueDate[] = $currentDate >= $dueDate[1] && $currentDate <= $dueDate[2];
-                $dueDate[] = $this->type;
-                return $dueDate;
-            }
-        }
-
-        return [];
-    }
-
-    /**
      * Récupère la déclaration sociale courante
-     * @param bool $forceCurrent
-     * @return Declaration|mixed
-     * @throws \Exception
+     * @throws Exception
      */
-    public function getDeclarations($forceCurrent = false)
+    public function getDeclarations(bool $forceCurrent = false): ?Declaration
     {
-        if ($this->type === Declaration::TYPE_SOCIAL) {
-            list ($annualyPeriod, $quarterlyPeriod)
-                = $this->isDueMonth() && !$forceCurrent ? $this->periodService->getPreviousPeriod() : $this->periodService->getCurrentPeriod();
+        $quarterlyPeriod = null;
+        if ($this->type === DeclarationTypeEnum::Social) {
+            [$annualyPeriod, $quarterlyPeriod]
+                = $this->isDueMonth() && !$forceCurrent ? $this->periodService->getPreviousPeriod(
+            ) : $this->periodService->getCurrentPeriod();
         } else {
-            list ($annualyPeriod) = $this->periodService->getCurrentPeriod();
+            [$annualyPeriod] = $this->periodService->getCurrentPeriod();
         }
 
-        if ($this->type === Declaration::TYPE_SOCIAL) {
+        if ($this->type === DeclarationTypeEnum::Social) {
             $declaration = $this->declarationRepository->getByDate(
                 $this->type,
                 $quarterlyPeriod->getYear(),
@@ -234,7 +184,7 @@ class DeclarationTypeService
             $declaration->setRevenue(0);
             $declaration->setTax(0);
 
-            if ($this->type === Declaration::TYPE_SOCIAL) {
+            if ($this->type === DeclarationTypeEnum::Social) {
                 $declaration->setPeriod($quarterlyPeriod);
             } else {
                 $declaration->setPeriod($annualyPeriod);
@@ -246,5 +196,22 @@ class DeclarationTypeService
         $this->entityManager->flush();
 
         return $declaration;
+    }
+
+    /**
+     * Verifie si on est dans un mois de déclaration
+     */
+    public function isDueMonth(): bool
+    {
+        $date = new DateTime();
+        $dueDatesMonth = $this->getDueMonth();
+
+        foreach ($dueDatesMonth as $index => $dueDateMonth) {
+            if (intval($date->format('m')) === $dueDateMonth) {
+                return true;
+            }
+        }
+
+        return false;
     }
 }
