@@ -2,42 +2,61 @@
 
 namespace App\Service;
 
-
 use App\Entity\Activity;
+use App\Entity\Company;
+use App\Repository\ActivityRepository;
+use App\Repository\ExperienceRepository;
+use App\Repository\InvoiceRepository;
 use DateInterval;
+use DateTime;
+use Doctrine\ORM\EntityManagerInterface;
+use Exception;
+use Symfony\Contracts\Translation\TranslatorInterface;
 
 class ReportService
 {
+    public function __construct(
+        private EntityManagerInterface $entityManager,
+        private InvoiceRepository      $invoiceRepository,
+        private ActivityRepository     $activityRepository,
+        private ExperienceRepository   $experienceRepository,
+        private TranslatorInterface    $translator,
+    ) {
+
+    }
+
     /**
      * Génère le tableau à partir d'un date donné
-     * @param \DateTime $currentDate
-     * @param array $activities
      * @param null $company
-     * @return array
-     * @throws \Exception
+     * @throws Exception
      */
-    public function generateMonth(\DateTime $currentDate, array $activities, $company = null): array
+    public function generateMonth(DateTime $currentDate, array $activities, $company = null): array
     {
         $activitiesData = [];
+        $now = (new DateTime())->format('Ymd');
 
-        for($i = 1; $i < $currentDate->format('N'); $i++) {
+        for ($i = 1; $i < $currentDate->format('N'); $i++) {
             $activitiesData[] = [
                 'selected' => false,
-                'date' => null,
-                'value' => null
+                'date'     => null,
+                'value'    => null
             ];
         }
 
         $emptyDays = $currentDate->format('t');
         for ($i = 1; $i <= $emptyDays; $i++) {
             $date = $currentDate->format('Ymd');
+            $dayInWeek = intval($currentDate->format('w'));
 
-            $activitiesData[$date] = [
-                'selected' => false,
-                'date' => clone $currentDate,
-                'value' => 1,
-                'company' => $company
-            ];
+            if ($dayInWeek !== 6 && $dayInWeek !== 0) {
+                $activitiesData[$date] = [
+                    'selected' => false,
+                    'date'     => clone $currentDate,
+                    'value'    => 1,
+                    'company'  => $company,
+                    'current'  => $date === $now
+                ];
+            }
             $currentDate->add(new DateInterval('P1D'));
         }
 
@@ -53,12 +72,11 @@ class ReportService
 
         $count = count($activitiesData);
 
-        for($i = $count; $i < ceil($count / 7) * 7; $i++)
-        {
+        for ($i = $count; $i < ceil($count / 7) * 7; $i++) {
             $activitiesData[] = [
                 'selected' => false,
-                'date' => null,
-                'value' => null
+                'date'     => null,
+                'value'    => null
             ];
         }
 
@@ -66,13 +84,81 @@ class ReportService
     }
 
     /**
-     * Envoi d'un mail à la fin du mois si une mission est en cours pour penser à envoyer un CRA
-     * @return array
+     * @throws Exception
      */
-    public function getNotifications()
-    {
-        $notifications = [];
+    public function getDashboard(array $viewData, DateTime $currentDate, int $year, int $month, ?Company $company
+    ): array {
+        $viewData['daysCount'] = $currentDate->format('t');
+        $viewData['years'] = $this->invoiceRepository->findYears();
 
-        return $notifications;
+        if (!in_array($viewData['activeYear'], $viewData['years'])) {
+            $viewData['years'][] = $viewData['activeYear'];
+        }
+        if (!in_array($viewData['currentYear'], $viewData['years'])) {
+            $viewData['years'][] = $viewData['currentYear'];
+        }
+
+        $viewData['months'] = [];
+        for ($i = 1; $i <= 12; $i++) {
+            $monthDate = new DateTime($viewData['activeYear'] . ($i < 10 ? '0' : '') . $i . '01');
+            $viewData['months'][] = [
+                'int' => $i,
+                'str' => $this->translator->trans($monthDate->format('F'))
+            ];
+        }
+
+        $viewData['companies'] = [];
+        $currentExperiences = $this->experienceRepository->getCurrents();
+        foreach ($currentExperiences as $experience) {
+            $viewData['companies'][] = $experience->getClient() ? $experience->getClient() : $experience->getCompany();
+        }
+        $viewData['activeCompany'] = count(
+                $viewData['companies']
+            ) == 1 ?? !$company ? $viewData['companies'][0] : $company;
+
+        $viewData['invoices'] = $this->invoiceRepository->getByDate($currentDate);
+
+        $viewData['companyActivities'] =
+            $viewData['activeCompany']
+                ? $this->activityRepository->findByCompanyAndDate($viewData['activeCompany'], $currentDate)
+                : $this->activityRepository->findByDate($currentDate);
+
+        return $viewData;
+    }
+
+    public function sendActivities(array $formData, DateTime $currentDate): void
+    {
+        $dayCount = 0;
+
+        foreach ($formData['activities'] as $activityData) {
+            if ($activityData['date'] && $activityData['selected']) {
+                $dayCount++;
+            }
+        }
+
+        $this->activityRepository->cleanByDate($currentDate);
+
+        foreach ($formData['activities'] as $activityData) {
+            if ($activityData['date'] && $activityData['selected']) {
+                $activity = new Activity();
+                $activity->setDate($activityData['date']);
+                $activity->setValue($activityData['value']);
+                $activity->setCompany($activityData['company']);
+
+                $this->entityManager->persist($activity);
+            }
+
+        }
+
+        $this->entityManager->flush();
+    }
+
+    /**
+     * @TODO
+     * Envoi d'un mail à la fin du mois si une mission est en cours pour penser à envoyer un CRA
+     */
+    public function getNotifications(): array
+    {
+        return [];
     }
 }
