@@ -26,17 +26,9 @@ Encore
     .addEntry('app', './assets/app.js')
     .addEntry('bootstrap', './assets/bootstrap.js')
 
-    .addEntry('recipes/main', './assets/js/recipes.js')
-    .addEntry('recipe/main', './assets/js/recipe.js')
-
-    .addStyleEntry('css/app', './assets/css/app.scss')
+    .addStyleEntry('css/admin', './assets/css/admin.scss')
     .addStyleEntry('css/login', './assets/css/login.scss')
     .addStyleEntry('css/index', './assets/css/index.scss')
-    .addStyleEntry('css/admin', './assets/css/admin.scss')
-    .addStyleEntry('css/dashboard', './assets/css/dashboard.scss')
-    .addStyleEntry('css/report', './assets/css/report.scss')
-    .addStyleEntry('css/report_pdf', './assets/css/report_pdf.scss')
-    .addStyleEntry('css/timeline', './assets/css/timeline.scss')
 
     // enables the Symfony UX Stimulus bridge (used in assets/bootstrap.js)
     .enableStimulusBridge('./assets/controllers.json')
@@ -75,7 +67,7 @@ Encore
     .enableSassLoader()
 
     // uncomment if you use TypeScript
-    .enableTypeScriptLoader()
+    //.enableTypeScriptLoader()
 
     // uncomment if you use React
     //.enableReactPreset()
@@ -87,7 +79,7 @@ Encore
     // uncomment if you're having problems with a jQuery plugin
     //.autoProvidejQuery()
 
-    .enableVueLoader()
+    //.enableVueLoader()
 ;
 
 module.exports = Encore.getWebpackConfig();
